Address remaining review comments on PR #928
- Move KeystorePrivateKey to zhtp/src/keystore_names.rs as single source of truth
- Update did_startup.rs to import from keystore_names module
- Update web4_utils.rs to re-export from zhtp::keystore_names
- Add 2x multiplier threshold check to fee calculation in domain.rs
- Simplify redundant OR condition in dao prefix check
