+++ zhtp/src/keystore_names.rs
@@ -1,13 +1,28 @@
-//! Centralized file naming constants for keystore and identity files
+//! Centralized file naming constants and types for keystore and identity files
 //!
-//! This module defines all standard filenames used in the keystore directory
-//! to avoid naming inconsistencies across the codebase.
+//! This module defines all standard filenames and data structures used in the keystore directory
+//! to avoid naming inconsistencies and type duplication across the codebase.
 //!
 //! Usage:
 //! ```ignore
 //! use crate::keystore_names::*;
 //! let path = keystore_path.join(NODE_IDENTITY_FILENAME);
 //! ```
+
+use serde::{Serialize, Deserialize};
+
+/// Private key storage format for keystore files
+///
+/// This struct defines the JSON format used to persist private keys to disk.
+/// It is shared across zhtp, zhtp-cli, and other components to ensure consistency.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct KeystorePrivateKey {
+    pub dilithium_sk: Vec<u8>,
+    #[serde(default)]
+    pub dilithium_pk: Vec<u8>,  // Optional for backward compatibility with old keystores
+    pub kyber_sk: Vec<u8>,
+    pub master_seed: Vec<u8>,
+}
 
 /// Node identity file (node's DID and identity metadata)
 pub const NODE_IDENTITY_FILENAME: &str = "node_identity.json";
